+++ src/assets/platform/platForm.png
