--- src/assets/platform/platformMap.json
@@ -1,50 +1,0 @@
-{
-  "frames": {
-    "green": {
-      "frame": {
-        "x": 0,
-        "y": 0,
-        "w": 100,
-        "h": 100
-      },
-      "spriteSourceSize": { "x": 0, "y": 0, "w": 100, "h": 100 },
-      "sourceSize": { "w": 100, "h": 100 }
-    },
-    "yellow": {
-      "frame": {
-        "x": 100,
-        "y": 0,
-        "w": 100,
-        "h": 100
-      },
-      "spriteSourceSize": { "x": 0, "y": 0, "w": 100, "h": 100 },
-      "sourceSize": { "w": 100, "h": 100 }
-    },
-    "red": {
-      "frame": {
-        "x": 200,
-        "y": 0,
-        "w": 100,
-        "h": 100
-      },
-      "spriteSourceSize": { "x": 0, "y": 0, "w": 100, "h": 100 },
-      "so,urceSize": { "w": 100, "h": 100 }
-    },
-    "purple": {
-      "frame": {
-        "x": 300,
-        "y": 0,
-        "w": 100,
-        "h": 100
-      },
-      "spriteSourceSize": { "x": 0, "y": 0, "w": 100, "h": 100 },
-      "sourceSize": { "w": 100, "h": 100 }
-    }
-  },
-  "meta": {
-    "image": "balloonSpritesheet.png",
-    "format": "RGBA8888",
-    "size": { "w": 400, "h": 100 },
-    "scale": "1"
-  }
-}
