+++ src/game/components/ScoreBoard.ts
@@ -1,5 +1,5 @@
 import { Container, BitmapText } from 'pixi.js'
-import config from '../../../../gameConfig.ts'
+import config from '../../../gameConfig.ts'
 
 export default class ScoreBoard {
   _displayedPlayerScore = 0
